Fix a shitload of warnings in 64 bits mode
Former-commit-id: c7792a7d5b1f85ab934da25324831b9daa3f47ff [formerly 3469974c48432be1f65808bff0ea39d9e22f5b50]
Former-commit-id: 5080815e9e1a3aebe237ff9a291b908ce0292eca

+++ include/Nazara/Utility/PixelFormat.inl
@@ -100,7 +100,7 @@ namespace Nz
 		counter |= blueMask;
 		counter |= alphaMask;
 
-		bitsPerPixel = counter.Count();
+		bitsPerPixel = static_cast<UInt8>(counter.Count());
 	}
 
 	inline bool PixelFormatInfo::Validate() const
@@ -116,7 +116,7 @@ namespace Nz
 
 		for (unsigned int i = 0; i < 4; ++i)
 		{
-			unsigned int usedBits = masks[i]->Count();
+			UInt8 usedBits = static_cast<UInt8>(masks[i]->Count());
 			if (usedBits == 0)
 				continue;
